add auth middleware

+++ server/src/main/java/com/ymonnier/restful/littleapp/controllers/UserController.java
@@ -1,5 +1,6 @@
 package com.ymonnier.restful.littleapp.controllers;
 
+import com.ymonnier.restful.littleapp.controllers.Auth.AuthorizationService;
 import com.ymonnier.restful.littleapp.models.User;
 import com.ymonnier.restful.littleapp.utilities.HibernateUtil;
 import com.ymonnier.restful.littleapp.utilities.errors.Error;
@@ -10,6 +11,7 @@ import javax.validation.Validation;
 import javax.validation.Validator;
 import javax.ws.rs.*;
 import javax.ws.rs.core.*;
+import java.net.Authenticator;
 import java.net.URI;
 import java.util.List;
 import java.util.Set;
@@ -25,18 +27,23 @@ import java.util.logging.Logger;
  */
 @Path("users/")
 @Produces(MediaType.APPLICATION_JSON)
-public class UserController {
+public class UserController extends AuthorizationService {
     private final static Logger LOGGER = Logger.getLogger(UserController.class.getSimpleName());
     public final static String PATH = "persons/";
+
+    public UserController(@Context HttpHeaders headers) {
+        super(headers);
+        LOGGER.info("#UserController... ");
+    }
 
     @Context
     UriInfo uriInfo;
 
     Validator validator = Validation.buildDefaultValidatorFactory().getValidator();
 
-
     /**
      * Allows to get a list of person.
+     *
      * @return response which contains the list
      * of person or errors.
      */
@@ -44,75 +51,36 @@ public class UserController {
     public Response index() {
         LOGGER.info("#GET ");
         Response response = null;
-        Session session = HibernateUtil.getSessionFactory().openSession();
-        try {
-            session.beginTransaction();
-            List persons = session.createQuery("from User")
-                    .list();
-            session.getTransaction().commit();
-            session.close();
-
-            response = Response
-                    .ok(persons)
-                    .build();
-        } catch (Exception exception) {
-            LOGGER.warning("#GET " + exception.getLocalizedMessage());
-            response = Error.internalServer(exception)
-                    .getResponse();
-            session.getTransaction().rollback();
-        }
-
-        return response;
-    }
-
-    /**
-     * Method handling HTTP POST requests. The returned object will be sent
-     * to the client as JSON Object.
-     *
-     * @return JSON Object created by the client.
-     */
-    @POST
-    @Consumes(MediaType.APPLICATION_JSON)
-    public Response create(User user) {
-        LOGGER.info("#POST " + user.toString());
-        Response response = null;
-        Set<ConstraintViolation<Object>> constraintViolations =
-                validator.validate(user);
-
-        if (constraintViolations.size() > 0) {
-            response = Error.badRequest(constraintViolations)
-                    .getResponse();
-            LOGGER.warning("#POST " + response.toString());
-        } else {
+        if (this.authorization) {
             Session session = HibernateUtil.getSessionFactory().openSession();
             try {
                 session.beginTransaction();
-                session.save(user);
+                List persons = session.createQuery("from User")
+                        .list();
                 session.getTransaction().commit();
-
-                URI builder = uriInfo.getAbsolutePathBuilder()
-                        //.path(String.valueOf(p.getId()))
-                        .build();
+                session.close();
 
                 response = Response
-                        .created(builder)
-                        .entity(user)
+                        .ok(persons)
                         .build();
-                LOGGER.info("#POST " + response.toString());
             } catch (Exception exception) {
-                LOGGER.warning("#POST " + exception.getLocalizedMessage());
+                LOGGER.warning("#GET " + exception.getLocalizedMessage());
                 response = Error.internalServer(exception)
                         .getResponse();
                 session.getTransaction().rollback();
             }
+        } else {
+            response = this.getUnauthorization();
         }
 
         return response;
     }
 
+
     /**
      * Http requet which get a specific
      * user depending on the id.
+     *
      * @param id user id
      * @return a response which contains the user or errors.
      */
@@ -121,29 +89,33 @@ public class UserController {
     public Response show(@PathParam("id") Long id) {
         LOGGER.info("#GET {" + id + "} ");
         Response response = null;
-        Session session = HibernateUtil.getSessionFactory().openSession();
-        try {
-            session.beginTransaction();
-            User user = session.get(User.class, id);
-            session.getTransaction().commit();
-            session.close();
+        if (this.authorization) {
+            Session session = HibernateUtil.getSessionFactory().openSession();
+            try {
+                session.beginTransaction();
+                User user = session.get(User.class, id);
+                session.getTransaction().commit();
+                session.close();
 
-            if (user == null) {
-                response = Error.notFound(String.valueOf(id))
+                if (user == null) {
+                    response = Error.notFound(String.valueOf(id))
+                            .getResponse();
+                    LOGGER.warning("#GET {" + id + "} " + response.toString());
+                } else {
+                    response = Response
+                            .ok(user)
+                            .build();
+                    LOGGER.info("#GET {" + id + "} " + response.toString());
+                }
+
+            } catch (Exception exception) {
+                response = Error.internalServer(exception)
                         .getResponse();
-                LOGGER.warning("#GET {" + id + "} " + response.toString());
-            } else {
-                response = Response
-                        .ok(user)
-                        .build();
-                LOGGER.info("#GET {" + id + "} " + response.toString());
+                LOGGER.warning("#GET {" + id + "} " + exception.getLocalizedMessage());
+                session.getTransaction().rollback();
             }
-
-        } catch (Exception exception) {
-            response = Error.internalServer(exception)
-                    .getResponse();
-            LOGGER.warning("#GET {" + id + "} " + exception.getLocalizedMessage());
-            session.getTransaction().rollback();
+        } else {
+            response = this.getUnauthorization();
         }
         return response;
     }
@@ -151,7 +123,8 @@ public class UserController {
 
     /**
      * Update HTTP Request which allows to update a user.
-     * @param id user id
+     *
+     * @param id   user id
      * @param user user data
      * @return a response which contains the user updated or errors.
      */
@@ -161,47 +134,52 @@ public class UserController {
     public Response update(@PathParam("id") Long id, User user) {
         LOGGER.info("#PUT {" + id + "} ");
         Response response = null;
+        if (this.authorization) {
+            Set<ConstraintViolation<Object>> constraintViolations =
+                    validator.validate(user);
 
-        Set<ConstraintViolation<Object>> constraintViolations =
-                validator.validate(user);
-
-        if (constraintViolations.size() > 0) {
-            response = Error.badRequest(constraintViolations)
-                    .getResponse();
-            LOGGER.warning("#PUT " + response.toString());
-        } else {
-            Session session = HibernateUtil.getSessionFactory().openSession();
-            try {
-                session.beginTransaction();
-                User userChecked = session.get(User.class, id);
-
-                if (userChecked == null) {
-                    response = Error.notFound(String.valueOf(id))
-                            .getResponse();
-                    LOGGER.warning("#PUT {" + id + "} " + response.toString());
-                } else {
-                    userChecked.update(user);
-                    session.update(userChecked);
-                    response = Response
-                            .ok(user)
-                            .build();
-                    LOGGER.info("#PUT {" + id + "} " + response.toString());
-                }
-
-                session.getTransaction().commit();
-                session.close();
-            } catch (Exception exception) {
-                response = Error.internalServer(exception)
+            if (constraintViolations.size() > 0) {
+                response = Error.badRequest(constraintViolations)
                         .getResponse();
-                LOGGER.warning("#PUT {" + id + "} " + exception.getLocalizedMessage());
-                session.getTransaction().rollback();
+                LOGGER.warning("#PUT " + response.toString());
+            } else {
+                Session session = HibernateUtil.getSessionFactory().openSession();
+                try {
+                    session.beginTransaction();
+                    User userChecked = session.get(User.class, id);
+
+                    if (userChecked == null) {
+                        response = Error.notFound(String.valueOf(id))
+                                .getResponse();
+                        LOGGER.warning("#PUT {" + id + "} " + response.toString());
+                    } else {
+                        userChecked.update(user);
+                        session.update(userChecked);
+                        response = Response
+                                .ok(user)
+                                .build();
+                        LOGGER.info("#PUT {" + id + "} " + response.toString());
+                    }
+
+                    session.getTransaction().commit();
+                    session.close();
+                } catch (Exception exception) {
+                    response = Error.internalServer(exception)
+                            .getResponse();
+                    LOGGER.warning("#PUT {" + id + "} " + exception.getLocalizedMessage());
+                    session.getTransaction().rollback();
+                }
             }
+        } else {
+            response = this.getUnauthorization();
         }
 
         return response;
     }
+
     /**
      * HTTP Request which allows to destroy a specific user.
+     *
      * @param id user id
      * @return a response which contains the user updated or errors.
      */
@@ -211,29 +189,33 @@ public class UserController {
     public Response destroy(@PathParam("id") Long id) {
         Response response = null;
 
-        Session session = HibernateUtil.getSessionFactory().openSession();
-        try {
-            session.beginTransaction();
-            User user = session.get(User.class, id);
+        if (this.authorization) {
+            Session session = HibernateUtil.getSessionFactory().openSession();
+            try {
+                session.beginTransaction();
+                User user = session.get(User.class, id);
 
-            if (user == null) {
-                response = Error.notFound(String.valueOf(id))
+                if (user == null) {
+                    response = Error.notFound(String.valueOf(id))
+                            .getResponse();
+                    LOGGER.warning("#GET {" + id + "} " + response.toString());
+                } else {
+                    session.delete(user);
+                    response = Response
+                            .ok(user)
+                            .build();
+                    LOGGER.info("#GET {" + id + "} " + response.toString());
+                }
+                session.getTransaction().commit();
+                session.close();
+            } catch (Exception exception) {
+                response = Error.internalServer(exception)
                         .getResponse();
-                LOGGER.warning("#GET {" + id + "} " + response.toString());
-            } else {
-                session.delete(user);
-                response = Response
-                        .ok(user)
-                        .build();
-                LOGGER.info("#GET {" + id + "} " + response.toString());
+                LOGGER.warning("#GET {" + id + "} " + exception.getLocalizedMessage());
+                session.getTransaction().rollback();
             }
-            session.getTransaction().commit();
-            session.close();
-        } catch (Exception exception) {
-            response = Error.internalServer(exception)
-                    .getResponse();
-            LOGGER.warning("#GET {" + id + "} " + exception.getLocalizedMessage());
-            session.getTransaction().rollback();
+        } else {
+            response = this.getUnauthorization();
         }
 
         return response;
